code reorg. virtual workspace. makefile
