+++ agent/Cargo.toml
@@ -1,9 +1,7 @@
 [package]
-name = "client"
+name = "agent"
 version = "0.1.0"
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 tokio = { version = "1.24.0", features = ["full"] }
