+++ agent/src/command/mod.rs
