+++ agent/src/common/mod.rs
@@ -1,0 +1,4 @@
+pub mod config;
+pub mod handlers;
+pub mod macros;
+pub mod rpc;
