+++ agent/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod command;
+pub mod common;
