--- command/src/lib.rs
@@ -1,1 +1,0 @@
-
