+++ rust-toolchain.toml
@@ -1,0 +1,4 @@
+[toolchain]
+channel = "nightly-2022-10-03"
+components = ["clippy", "llvm-tools-preview", "rustfmt"]
+targets = ["wasm32-unknown-unknown"]
