+++ utils/jsonrpc/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "jsonrpc"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.68"
+async-channel = "1.8.0"
+async-trait = "0.1.61"
+futures-util = { version = "0.3", default-features = false, features = ["sink", "std"] }
+log = "0.4.17"
+reqwest = { version = "0.11.13", features = ["json"] }
+serde_json = "1.0.91"
+tokio = { version = "1.16", features = ["full"] }
+tokio-tungstenite = { version = "0.18.0", features = ["native-tls"] }
+url = "2.3.1"
